requirements.txt
gunicorn
flask
flask-cors
numpy
pillow
tensorflow
